Tests/LibWeb/Text/input/link-element-media-change-off.html

<!DOCTYPE html>
<html>
    <head>
        <link rel="stylesheet" href="body-background-color-red.css" media="all" onload="this.media='none'">
    </head>
    <body>
        <script src="include.js"></script>
        <script>
            test(() => {
                window.onload = function () {
                    println("document background: " + getComputedStyle(document.body).backgroundColor);
                };
            });
        </script>
    </body>
</html>
