Tests/LibWeb/Text/expected/link-element-media-change-off.txt

document background: rgba(0, 0, 0, 0)
